fix send notificate

+++ src/service/action/AuthenticateAction.js
@@ -6,24 +6,26 @@ import { getRole } from "../util/util"
 
 export const checkSession = () => {
     var token = localStorage.getItem('token')
-    var decode = jwtDecode(token)
-    var time = null
+    if (token !== null) {
+        var decode = jwtDecode(token)
+        var time = null
 
-    Object.keys(decode).forEach(key => {
-        let res = key.split('/')
-        if (res[res.length - 1] === 'exp') {
-            time = moment.unix(decode[key]).format("DD/MM/YYYY HH:mm:ss")
-        }
-    })
-    var now = moment().format("DD/MM/YYYY HH:mm:ss")
-    var diff = moment(time, "DD/MM/YYYY HH:mm:ss").diff(moment(now, "DD/MM/YYYY HH:mm:ss"))
-    return (dispatch) => {
-        if (diff <= 0 || getRole() === 'admin') {
-            localStorage.clear()
-            dispatch(sessionTimeOut())
-            history.push('/login')
-        } else {
-            dispatch(sessionAllow())
+        Object.keys(decode).forEach(key => {
+            let res = key.split('/')
+            if (res[res.length - 1] === 'exp') {
+                time = moment.unix(decode[key]).format("DD/MM/YYYY HH:mm:ss")
+            }
+        })
+        var now = moment().format("DD/MM/YYYY HH:mm:ss")
+        var diff = moment(time, "DD/MM/YYYY HH:mm:ss").diff(moment(now, "DD/MM/YYYY HH:mm:ss"))
+        return (dispatch) => {
+            if (diff <= 0 || getRole() === 'admin') {
+                localStorage.clear()
+                dispatch(sessionTimeOut())
+                history.push('/login')
+            } else {
+                dispatch(sessionAllow())
+            }
         }
     }
 }
